+++ asm/custom-functions/src/lib.rs
@@ -35,6 +35,7 @@ macro_rules! println {
 }
 
 #[panic_handler]
-fn panic(_: &core::panic::PanicInfo) -> ! {
+fn panic(info: &core::panic::PanicInfo) -> ! {
+    println!("{}", info);
     loop {}
 }
