+++ asm/patch_diffs/custom_items_diff.txt
@@ -1,6 +1,6 @@
 main.dol:
   0x80251D04:
-    Data: [0x7F, 0xC3, 0xF3, 0x78, 0x48, 0x42, 0x6B, 0xAD, 0x48, 0x00, 0x00, 0x3C]
+    Data: [0x7F, 0xC3, 0xF3, 0x78, 0x48, 0x42, 0x84, 0x31, 0x48, 0x00, 0x00, 0x3C]
   0x80253CF8:
     Data: [0x48, 0x42, 0x46, 0x49]
   0x804E8F40:
@@ -14,7 +14,7 @@ main.dol:
       0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10,
       0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00]
   0x8016FB00:
-    Data: [0x7C, 0xB9, 0x2B, 0x78, 0x7F, 0x45, 0xD3, 0x78, 0x48, 0x50, 0x92, 0x09,
+    Data: [0x7C, 0xB9, 0x2B, 0x78, 0x7F, 0x45, 0xD3, 0x78, 0x48, 0x50, 0xAA, 0xE9,
       0x7F, 0x25, 0xCB, 0x78, 0x90, 0x61, 0x00, 0x0C, 0x2C, 0x03, 0x00, 0x00, 0x40,
       0x82, 0x00, 0x0C]
   0x8016FC38:
