+++ asm/patch_diffs/ss_necessary_diff.txt
@@ -3,9 +3,9 @@ main.dol:
     Data: [0x80, 0x1C, 0x00, 0x04, 0x54, 0x00, 0xE7, 0xBE, 0x98, 0x1C, 0x12, 0x09,
       0x48, 0x00, 0x00, 0x18]
   0x80115CF8:
-    Data: [0x48, 0x56, 0x2F, 0x81]
+    Data: [0x48, 0x56, 0x48, 0x61]
   0x80115F98:
-    Data: [0x48, 0x56, 0x2C, 0xE1]
+    Data: [0x48, 0x56, 0x45, 0xC1]
   0x800A0968:
     Data: [0x38, 0x60, 0x00, 0x01, 0x48, 0x00, 0x00, 0x34]
   0x80115A04:
@@ -21,7 +21,7 @@ main.dol:
   0x800B2774:
     Data: [0xA0, 0x7A, 0x14, 0x7A, 0x38, 0x03, 0x00, 0x01, 0xB0, 0x1A, 0x14, 0x7A]
   0x801D45EC:
-    Data: [0x48, 0x4A, 0x46, 0xCD]
+    Data: [0x48, 0x4A, 0x5F, 0xAD]
   0x801D4B20:
     Data: [0x38, 0xA0, 0xFF, 0xFF]
   0x8005E2F0:
@@ -87,7 +87,7 @@ main.dol:
     Data: [0x60, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00,
       0x60, 0x00, 0x00, 0x00]
   0x801BB9BC:
-    Data: [0x48, 0x4B, 0xCC, 0xD0]
+    Data: [0x48, 0x4B, 0xE5, 0x54]
   0x801B9A10:
     Data: [0x38, 0xA0, 0x00, 0x02]
   0x801B9A44:
@@ -123,11 +123,11 @@ main.dol:
   0x801AFF2C:
     Data: [0x41, 0x81, 0x08, 0x5C]
   0x801B0788:
-    Data: [0x48, 0x4C, 0x81, 0xD9, 0x4B, 0xFF, 0xFF, 0xD8]
+    Data: [0x48, 0x4C, 0x9A, 0x5D, 0x4B, 0xFF, 0xFF, 0xD8]
   0x801B19E0:
     Data: [0x38, 0x00, 0x00, 0x00]
   0x801BB980:
-    Data: [0x48, 0x4B, 0xD5, 0x99]
+    Data: [0x48, 0x4B, 0xEB, 0x7D]
   0x804DBA00:
     Data: [0x00, 0x00]
   0x8025685C:
@@ -151,18 +151,18 @@ main.dol:
   0x802D7660:
     Data: [0x48, 0x3A, 0x0E, 0x6D]
   0x8024CFFC:
-    Data: [0xA0, 0x7D, 0x03, 0x30, 0x48, 0x42, 0xC2, 0xC1, 0x2C, 0x03, 0x00, 0x04]
+    Data: [0xA0, 0x7D, 0x03, 0x30, 0x48, 0x42, 0xDB, 0x45, 0x2C, 0x03, 0x00, 0x04]
   0x8024D00C:
     Data: [0x60, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00]
   0x801AFF50:
     Data: [0x28, 0x04, 0xFF, 0xFF, 0x41, 0x82, 0x00, 0x0C, 0x7C, 0x83, 0x23, 0x78,
-      0x48, 0x4C, 0xB5, 0xB9]
+      0x48, 0x4C, 0x89, 0x69]
   0x8006358C:
-    Data: [0x48, 0x61, 0x59, 0xE9]
+    Data: [0x48, 0x61, 0x72, 0x6D]
   0x800A5C88:
     Data: [0x48, 0x5D, 0x29, 0xF4]
   0x801BA668:
-    Data: [0x48, 0x4B, 0xEC, 0x30]
+    Data: [0x48, 0x4C, 0x04, 0xB4]
   0x8016C718:
     Data: [0x38, 0x60, 0x00, 0x01]
   0x8016C808:
@@ -174,47 +174,47 @@ main.dol:
   0x804E7CD8:
     Data: [0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x23]
   0x8005E288:
-    Data: [0x48, 0x61, 0xB8, 0x79]
+    Data: [0x48, 0x61, 0xD1, 0x9D]
   0x8005E2AC:
-    Data: [0x48, 0x61, 0xB8, 0x55]
+    Data: [0x48, 0x61, 0xD1, 0x79]
   0x801C7A3C:
-    Data: [0x48, 0x4B, 0x20, 0xC5]
+    Data: [0x48, 0x4B, 0x39, 0xE9]
   0x801C7B70:
-    Data: [0x48, 0x4B, 0x1F, 0x91]
+    Data: [0x48, 0x4B, 0x38, 0xB5]
   0x801CA16C:
-    Data: [0x48, 0x4A, 0xF9, 0x95]
+    Data: [0x48, 0x4B, 0x12, 0xB9]
   0x801E21E4:
-    Data: [0x48, 0x49, 0x79, 0x1D]
+    Data: [0x48, 0x49, 0x92, 0x41]
   0x801C5D8C:
-    Data: [0x48, 0x4B, 0x3D, 0x85]
+    Data: [0x48, 0x4B, 0x56, 0xA9]
   0x8024ACF8:
-    Data: [0x48, 0x42, 0xEE, 0x29]
+    Data: [0x48, 0x43, 0x07, 0x4D]
   0x800C7C50:
-    Data: [0x48, 0x5B, 0x1E, 0xD1]
+    Data: [0x48, 0x5B, 0x37, 0xF5]
   0x801BF4F8:
-    Data: [0x48, 0x4B, 0x9E, 0x2C]
+    Data: [0x48, 0x4B, 0xB6, 0xB0]
   0x8026A7D4:
     Data: [0x38, 0x9F, 0x02, 0x30]
   0x8026C3F4:
-    Data: [0xA0, 0x7F, 0x12, 0x00, 0x48, 0x40, 0xD2, 0x31]
+    Data: [0xA0, 0x7F, 0x12, 0x00, 0x48, 0x40, 0xEC, 0x75]
   0x8026C478:
-    Data: [0xA0, 0x7F, 0x12, 0x00, 0x48, 0x40, 0xD3, 0x11, 0x2C, 0x03, 0x00, 0x00,
+    Data: [0xA0, 0x7F, 0x12, 0x00, 0x48, 0x40, 0xEC, 0x35, 0x2C, 0x03, 0x00, 0x00,
       0x41, 0x82, 0x00, 0x20]
   0x8024D54C:
-    Data: [0xA0, 0x7D, 0x0D, 0x44, 0x48, 0x42, 0xC2, 0xE5, 0x83, 0xE1, 0x00, 0x1C,
+    Data: [0xA0, 0x7D, 0x0D, 0x44, 0x48, 0x42, 0xDD, 0xA9, 0x83, 0xE1, 0x00, 0x1C,
       0x83, 0xC1, 0x00, 0x18, 0x83, 0xA1, 0x00, 0x14, 0x80, 0x01, 0x00, 0x24, 0x7C,
       0x08, 0x03, 0xA6, 0x38, 0x21, 0x00, 0x20, 0x4E, 0x80, 0x00, 0x20]
   0x80541838:
-    Data: [0x80, 0x67, 0xB0, 0x8C]
+    Data: [0x80, 0x67, 0x8A, 0x48]
   0x8054184C:
-    Data: [0x80, 0x67, 0xB0, 0xD8]
+    Data: [0x80, 0x67, 0x8A, 0x94]
   0x80541854:
-    Data: [0x80, 0x2E, 0x15, 0xE0, 0x80, 0x67, 0xB2, 0x64]
+    Data: [0x80, 0x2E, 0x15, 0xE0, 0x80, 0x67, 0x8C, 0xF0]
   0x80541860:
-    Data: [0x80, 0x2E, 0x17, 0xA0, 0x80, 0x67, 0xB0, 0xFC, 0x80, 0x05, 0x08, 0x90,
+    Data: [0x80, 0x2E, 0x17, 0xA0, 0x80, 0x67, 0x8A, 0xC4, 0x80, 0x05, 0x08, 0x90,
       0x80, 0x05, 0x08, 0xF0]
   0x8054188C:
-    Data: [0x80, 0x67, 0xB2, 0x6C]
+    Data: [0x80, 0x67, 0x8D, 0x30]
   0x804E2788:
     Data: [0x00, 0x01, 0x00, 0x01]
 d_a_obj_time_door_beforeNP.rel:
